updated api, added ETSI OSM vimconn
